Pass PCB an absolute path to PCB when editing the layout.
Depending on the lrealpath implementation found at compile time in PCB,
the path returned by PCB may not be an absolute path (as we require)
if the file does not yet exist on disk. (e.g. a new project).

Passing a full path on PCB's command line ensures it will return an
absolute path when queried via D-Bus.