Rename translations to drop un-needed ISO-3166 country codes.
From Debian bug #336796:

Using a country part in a translation file is a discouraged practice
except in very few cases (such as pt_BR).

Using a fr_FR.po file instead of a fr.po file prevents users of fr_CA,
fr_BE, fr_LU, fr_CH and all other existing and future locales for
French to benefit from the French translation of the program.

In general PO files should only be named after the
ISO_639 code of the given language and should not use a country part
with a ISO-3166 code. The only accepted expcetions to this are:

-pt_BR for Brazilian Portuguese and pt alone for "standard Portuguese"
-zh_CN for "Simplified Chinese" use in mailand China and Singapore
-zh_TW for "Traditional Chinese" used in Taiwan